Move formatting functions to reducer, general refactor
